+++ src/test/java/de/gematik/demis/nrs/service/lookup/LookupTreeIT.java
@@ -52,7 +52,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest(
     webEnvironment = SpringBootTest.WebEnvironment.NONE,
-    properties = "nrs.lookup-data-directory=src/test/resources/lookup/maps")
+    properties = {
+      "nrs.lookup-data-directory=src/test/resources/lookup/maps",
+      "nrs.checkWorkaroundIsWorking=false"
+    })
 class LookupTreeIT {
 
   private static final Random r = new Random(1L);